// app/src/lib/shared/printers/models.generated.ts
export const UPSTREAM_PRINTERS = {
	tag: 'v02.08.02.61',
	commit: '926a7192574bcb9b3a732e1ec59a46d79cb45466'
} as const;